Move some code to a new file, writing a new implementation of the filter.
